🌱 Split oversized dashboard components
Refactor oversized React TypeScript components by extracting form, list, section, and hook/presentation subcomponents while preserving behavior.


# Conflicts:
#	web/src/components/cards/CardWrapper.tsx
#	web/src/components/cards/console-missions/ConsoleOfflineDetectionCard.tsx
#	web/src/components/cards/console-missions/useOfflineDetection.ts
